+++ src/emissionprobabilitycomputer.hpp
@@ -14,7 +14,7 @@
 * Computes the emission probabilities for a variant position.
 **/
 
-typedef std::vector<std::vector<long double>> ProbabilityMatrix;
+typedef std::vector<std::vector<double>> ProbabilityMatrix;
 
 class EmissionProbabilityComputer {
 public:
@@ -23,13 +23,13 @@ public:
 	 **/
 	EmissionProbabilityComputer(std::shared_ptr<UniqueKmers> uniquekmers, ProbabilityTable* probabilities);
 	/** get emission probability for a state in the HMM **/
-	long double get_emission_probability(unsigned short allele_id1, unsigned short allele_id2) const;
+	double get_emission_probability(unsigned short allele_id1, unsigned short allele_id2) const;
 
 private:
 	std::shared_ptr<UniqueKmers> uniquekmers;
 	ProbabilityTable* probabilities;
 	bool all_zeros;
 	ProbabilityMatrix state_to_prob;
-	long double compute_emission_probability(unsigned short allele1, unsigned short allele2, bool allele1_undefined, bool allele2_undefined);
+	double compute_emission_probability(unsigned short allele1, unsigned short allele2, bool allele1_undefined, bool allele2_undefined);
 };
 # endif // EMISSIONPROBABILITYCOMPUTER_H
